+++ ROOT_demo/Assets/Script/Board.cs
@@ -831,7 +831,7 @@ namespace ROOT
             }
         }
 
-        public int CalculateTotalIncomes()
+        public float CalculateTotalIncomes()
         {
             return Units.Where(unit => unit.IsSource).Sum(unit => unit.SignalCore.CalScore());
         }
